+++ intetests/audience_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestListInsights(t *testing.T) {
+func TestCRUDAudienceSuccess(t *testing.T) {
 	dom, teardownSuite := setupSuite(t)
 	defer teardownSuite(t)
 	ctx := context.Background()
@@ -19,123 +19,49 @@ func TestListInsights(t *testing.T) {
 		IsAdmin:  true,
 	})
 	assert.NoError(t, err)
-
-	user, err := dom.CreateUser(ctx, domain.User{
-		Username: "user",
-		Password: "password",
-		IsAdmin:  false,
-	})
+	asset, err := dom.AddAsset(ctx, admin, domain.Asset{
+		Data: &domain.Audience{
+			AgeMax:            30,
+			AgeMin:            20,
+			Gender:            domain.FemaleGenderType,
+			Country:           "Sweden",
+			HoursSpent:        3,
+			NumberOfPurchases: 3,
+			Description:       "bla bla",
+		}})
+	assert.NotNil(t, asset)
 	assert.NoError(t, err)
-	for i := 1; i <= 100; i++ {
-		desc := fmt.Sprintf("example %d", i)
-		asset, err := dom.AddAsset(ctx, admin, domain.Asset{
-			Data: &domain.Insight{
-				Text:        "40% of millenials spend more than 3hours on social media daily",
-				Description: desc,
-			}})
-		assert.NotNil(t, asset)
-		assert.NoError(t, err)
-		assert.Equal(t, uint(i), asset.ID)
-		assert.Equal(t, desc, asset.Data.(*domain.Insight).Description)
-	}
-	qa := domain.QueryAssets{
-		Limit:  10,
-		LastID: 0,
-		Type:   domain.InsightAssetType,
-		IsDesc: false,
-	}
-	la, err := dom.ListAssets(ctx, user, qa, nil)
-	fmt.Println(la)
+	assert.Equal(t, uint(1), asset.ID)
+	assert.Equal(t, "bla bla", asset.Data.(*domain.Audience).Description)
 
+	asset, err = dom.UpdateAsset(ctx, admin, domain.Asset{
+		ID: 1,
+		Data: &domain.Audience{
+			AgeMax:            30,
+			AgeMin:            20,
+			Gender:            domain.FemaleGenderType,
+			Country:           "Sweden",
+			HoursSpent:        3,
+			NumberOfPurchases: 3,
+			Description:       "bla bla 2",
+		}})
+	assert.NotNil(t, asset)
 	assert.NoError(t, err)
-	assert.NotNil(t, la)
-	assert.Equal(t, 10, len(la.Assets))
-	assert.Equal(t, uint(1), la.FirstID)
-	assert.Equal(t, uint(10), la.LastID)
+	assert.Equal(t, uint(1), asset.ID)
+	assert.Equal(t, "bla bla 2", asset.Data.(*domain.Audience).Description)
 
-	qa = domain.QueryAssets{
-		Limit:  10,
-		LastID: 101,
-		Type:   domain.InsightAssetType,
-		IsDesc: true,
-	}
-	la, err = dom.ListAssets(ctx, user, qa, nil)
-	fmt.Println(la)
-
+	gottenAsset, err := dom.GetAsset(ctx, admin, asset.ID, domain.AudienceAssetType)
+	assert.NotNil(t, gottenAsset)
 	assert.NoError(t, err)
-	assert.NotNil(t, la)
-	assert.Equal(t, 10, len(la.Assets))
-	assert.Equal(t, uint(100), la.FirstID)
-	assert.Equal(t, uint(91), la.LastID)
+	assert.EqualValues(t, asset, gottenAsset)
+	err = dom.DeleteAsset(ctx, admin, asset.ID, domain.AudienceAssetType)
+	assert.Nil(t, err)
+
+	_, err = dom.GetAsset(ctx, admin, asset.ID, domain.AudienceAssetType)
+	assert.NotNil(t, err)
 }
 
-func TestListCharts(t *testing.T) {
-	dom, teardownSuite := setupSuite(t)
-	defer teardownSuite(t)
-	ctx := context.Background()
-	admin, err := dom.CreateUser(ctx, domain.User{
-		Username: "admin",
-		Password: "password",
-		IsAdmin:  true,
-	})
-	assert.NoError(t, err)
-
-	user, err := dom.CreateUser(ctx, domain.User{
-		Username: "user",
-		Password: "password",
-		IsAdmin:  false,
-	})
-	assert.NoError(t, err)
-	for i := 1; i <= 100; i++ {
-		desc := fmt.Sprintf("example %d", i)
-		asset, err := dom.AddAsset(ctx, admin, domain.Asset{
-			Data: &domain.Chart{
-				Description: desc,
-				Title:       "Relationship between tax and GDP",
-				XTitle:      "GDP",
-				YTitle:      "Tax",
-				Data: domain.XYData{
-					X: []interface{}{1, 2, 3, 4, 5},
-					Y: []interface{}{1, 2, 3, 4, 5},
-				},
-			}})
-		assert.NotNil(t, asset)
-		assert.NoError(t, err)
-		assert.Equal(t, uint(i), asset.ID)
-		assert.Equal(t, desc, asset.Data.(*domain.Chart).Description)
-	}
-	qa := domain.QueryAssets{
-		Limit:  10,
-		LastID: 0,
-		Type:   domain.ChartAssetType,
-		IsDesc: false,
-	}
-	la, err := dom.ListAssets(ctx, user, qa, nil)
-	fmt.Println(la)
-
-	assert.NoError(t, err)
-	assert.NotNil(t, la)
-	assert.Equal(t, 10, len(la.Assets))
-	assert.Equal(t, uint(1), la.FirstID)
-	assert.Equal(t, uint(10), la.LastID)
-
-	qa = domain.QueryAssets{
-		Limit:  10,
-		LastID: 101,
-		Type:   domain.ChartAssetType,
-		IsDesc: true,
-	}
-	la, err = dom.ListAssets(ctx, user, qa, nil)
-	fmt.Println(la)
-
-	assert.NoError(t, err)
-	assert.NotNil(t, la)
-	assert.Equal(t, 10, len(la.Assets))
-	assert.Equal(t, uint(100), la.FirstID)
-	assert.Equal(t, uint(91), la.LastID)
-}
-
-func TestListAudiences(t *testing.T) {
+func TestListAudiencesSuccess(t *testing.T) {
 	dom, teardownSuite := setupSuite(t)
 	defer teardownSuite(t)
 	ctx := context.Background()
@@ -176,7 +102,6 @@ func TestListAudiences(t *testing.T) {
 		IsDesc: false,
 	}
 	la, err := dom.ListAssets(ctx, user, qa, nil)
-	fmt.Println(la)
 
 	assert.NoError(t, err)
 	assert.NotNil(t, la)
@@ -191,11 +116,101 @@ func TestListAudiences(t *testing.T) {
 		IsDesc: true,
 	}
 	la, err = dom.ListAssets(ctx, user, qa, nil)
-	fmt.Println(la)
 
 	assert.NoError(t, err)
 	assert.NotNil(t, la)
 	assert.Equal(t, 10, len(la.Assets))
 	assert.Equal(t, uint(100), la.FirstID)
 	assert.Equal(t, uint(91), la.LastID)
+}
+
+func TestListFavouriteAudiencesSuccess(t *testing.T) {
+	dom, teardownSuite := setupSuite(t)
+	defer teardownSuite(t)
+	ctx := context.Background()
+	admin, err := dom.CreateUser(ctx, domain.User{
+		Username: "admin",
+		Password: "password",
+		IsAdmin:  true,
+	})
+	assert.NoError(t, err)
+
+	user, err := dom.CreateUser(ctx, domain.User{
+		Username: "user",
+		Password: "password",
+		IsAdmin:  false,
+	})
+	assert.NoError(t, err)
+	for i := 1; i <= 100; i++ {
+		desc := fmt.Sprintf("example %d", i)
+		asset, err := dom.AddAsset(ctx, admin, domain.Asset{
+			Data: &domain.Audience{
+				AgeMax:            30,
+				AgeMin:            20,
+				Gender:            domain.FemaleGenderType,
+				Country:           "Sweden",
+				HoursSpent:        3,
+				NumberOfPurchases: 3,
+				Description:       desc,
+			}})
+		assert.NotNil(t, asset)
+		assert.NoError(t, err)
+		assert.Equal(t, uint(i), asset.ID)
+		assert.Equal(t, desc, asset.Data.(*domain.Audience).Description)
+		if i%2 == 0 {
+			err := dom.FavouriteAsset(ctx, user, asset.ID, domain.AudienceAssetType, true)
+			assert.NoError(t, err)
+		}
+	}
+	qa := domain.QueryAssets{
+		Limit:  10,
+		LastID: 0,
+		Type:   domain.AudienceAssetType,
+		IsDesc: false,
+	}
+	favQuery := domain.QueryFavouriteAssets{
+		FromUserID: user.ID,
+		OnlyFav:    true,
+	}
+	la, err := dom.ListAssets(ctx, user, qa, &favQuery)
+	assert.NoError(t, err)
+	assert.NotNil(t, la)
+	assert.Equal(t, 10, len(la.Assets))
+	assert.True(t, la.Assets[0].IsFavourite)
+	assert.Equal(t, uint(2), la.FirstID)
+	assert.Equal(t, uint(20), la.LastID)
+
+	qa = domain.QueryAssets{
+		Limit:  10,
+		LastID: 101,
+		Type:   domain.AudienceAssetType,
+		IsDesc: true,
+	}
+	la, err = dom.ListAssets(ctx, user, qa, &favQuery)
+	assert.NoError(t, err)
+	assert.NotNil(t, la)
+	assert.Equal(t, 10, len(la.Assets))
+	assert.True(t, la.Assets[0].IsFavourite)
+	assert.Equal(t, uint(100), la.FirstID)
+	assert.Equal(t, uint(82), la.LastID)
+
+	qa = domain.QueryAssets{
+		Limit:  10,
+		LastID: 0,
+		Type:   domain.AudienceAssetType,
+		IsDesc: false,
+	}
+	favQuery = domain.QueryFavouriteAssets{
+		FromUserID: user.ID,
+		OnlyFav:    false,
+	}
+
+	la, err = dom.ListAssets(ctx, user, qa, &favQuery)
+	assert.NoError(t, err)
+	assert.NotNil(t, la)
+	assert.Equal(t, 10, len(la.Assets))
+	assert.False(t, la.Assets[0].IsFavourite)
+	assert.True(t, la.Assets[1].IsFavourite)
+	assert.Equal(t, uint(1), la.FirstID)
+	assert.Equal(t, uint(10), la.LastID)
 }
